Reorganized all code into src/ directory
