+++ src/app.js
@@ -1,4 +1,4 @@
 const spawnClient = require("./client-spawner.js");
-const appConfig = require("./config.json");
+const appConfig = require("./../config.json");
 
 spawnClient(appConfig);
