+++ src/client-spawner.js
